refactor: 서비스 -> 컨트롤러로 책임 변경
